Started introducing what Asp.NET uses for database interaction The main framework that is used in order to interact with a database in .net is called Entity Framework Core.
That way that it works is that it is mapping to the database and you
interact with the context that is exposed throughout the code.
At first you must create a new class in the "Models" folder
Then you must generate a constructor that passes in DbContextOptions of
whatever type your Class is named, and also pass those same options to
the DbContext base class that you will be inheriting from.

After, you must create a connectionstring for the database inside of the
appsettings.json file

Then you dependency inject the DbContext as a services of type of your
DbContext class that you created inside of the programs.cs class. All of
You should also set the actual implementation as builder.Services so
that it can be used throughout the project

Inside each of the repositories, you should also use a constructor that
is ustilizing dependancy injection by passing in a DbContext class that
you created into the constructor for the class. You set it to inherit
from the interface you created and implement each of it's required
properties, now using the dbContext that you just injected.