Created a DbInitializer.cs
When creating a project that doesn't already have an existing database to call from, you can create a class for Initializing the data.

Inside the DbInitializer, you make it static because you won't be needing multiple instances of this class. Due to it being static though, you aren't able to use dependency injection, so in order to bring the necessary dbcontext to write to the db, you must bring it from the applicationbuilder. This application builder will be passed along when you call the initializer from the Program.cs. After bring in the context, you can start writing your data. You should wrap it in an if statement, so that if there is already data in the context, you don't overwrite it with initialization data. After inputting all the information that you want to initialize the database with, you have to use the context.SaveChanges() method in order to actually write the changes to the context, so that it gets pushed with the next database sync