Learned about routing and tag helpers and implemented them
Routing takes place by the maping routes that you have layed out in the Program.cs. ASP.NET is not able to map to a single file that contains all of the existing code, so it relies on the controller classes to route them through the correct action methods.

In the case that a request doesn't specify what controller or action method it should be routed through, you can layout default routes inside of the Program.cs file. The default is for the controller=Home/action=Index. You are able to add as many segments as you want/need. You are also able to restrict the type that a segment can validate.

Tag Helpers are a Asp.NET MVC class that help out regular HTML anchor tags. They have the ability to help the user so that they don't have to statically write down the links going to another page. With tag helpers, you are able to trigger server-side code by specifying the asp-controller and the as-action. Each one of those maps to the Controller file and the Action method within that file accordingly. In order to be able to utilize the class, you have to import it per file, or add it to the _ViewImports.cshtml file.