Learned about Sessions, View Components, & Tag Helpers
Sessions are the way that we can keep information going from one request on the server to another. Servers by default are stateless, which means that they don't hold any information from one request to another. The way that we are able to keep session information is through cookies. They tag a user's machine and help follow it from request to request. That functionality also has to be reflected within the model so that the same shopping instance can be followed throughout the requests. Sessions also need to be enabled through the services.AddSession() method and the app.UseSession() methods in the program.cs

View components are parts of a gui that can be used in multiple different locations. They are different to partial views because view components have their own code, as apposed to partial views getting the same model code passed from the parent views. This gives them very good functionality in that they can be used whenever a part of the gui needs to be dynamic in what it renders. View components have a view part and a class part. One of the ways to create the class part is by creating a viewcomponent class derived from the "ViewComponent" base class. View components don't necessarily have to return a view, they can just return a string or data if needed.To store them, you store the class file within a components folder in the project, and the component view within a component folder inside of the Views/Shared folders. The name for the views for viewcomponents is Default.cshtml. To call a view component you can either invoke it with "@await Component.InvokeAsync()" or by using a tag helper

Tag Helpers are very useful if you think that a component is going to be reused multiple times. Tag helpers are rendered server-side and are called through html. To create a tag helper class, you must inherit from the base tag helper class. Then you must add the tag helpers folder to the _ViewImports.cshtml file to be able to use it in the cshtml files.