Learned about forms, using model binding, validations, & razor pages
Forms are a way for the user to be able to submit information into the
application. Forms can be created using tag helpers in order to bind
to properties in the controller. If a controller is not specified, an
action method that fits will be looked for within the parent view's
controller. Within a control, use the "asp-for" tag helper to link that
item to the property in the model.

Model Binding is a process in Asp.NET that will extract data from the
request & provide that to the controller action. The information needed
for the model binding is pulled from data from the request, as well as
other sources such as the form. The information is the piped into the
parameters needed for the controller action.

Validation happen whenever we get information from the client to ensure
that we can work with the data. If the data is not usable, then we need
to display errors for the user. The model binders can provide validation
with the use of the "ModelState". Using the "IsValid()" method, we can
validate everything in one go. Using "GetValidationState()" validates a
single property. While default validation can be basic, we can add
custom validation. On top of that you can check more flags that need to
be checked using Property Attributes in the model. Some of the common
ones are: [Required], [StringLength], [Range], [RegularExpresion]. These
validations can be used for custom error messages. Doing the validation
just in the model however will only validated on the server side.

Adding client side validation can be done due to Asp.NET providing
metadata that third parties can hook into and do it on the client side.
An example of a library can be one from jquery.

Razor pages are an equivalent to MVC pages. Anything done in one can be
done in the other. The only differences is that with Razor pages you
don't need a controller class, making the creation process a bit
simpler. With Razor pages, you have a view(cshtml), and a code-behind
file. On any project, you can use both MVC & Razor at the same time.

One of the key differences in Razor is that the routing is different.
With Razor, it expects to live within the "Pages" folder within the root
directory of the project. Sites are then navigated to by filename.
Default get requests are handled by the "OnGet()" method, while post
requests are handled by the "OnPost()" methods. In order to bind
something to the a property in the model class of a razor page, you need
to use the "[BindProperty]" attribute.