Learned about unit testing in Asp.NET
Unit tests are automated pieces of code that the user writes in ordeter
to test a unit of work in the system. Unit tests are usually applied to
individual methods or classes, and most commonly testing the public
facing components. So in a controller, it would be the parts that are
naviagted to, or in a IRepository, the parts that would be interacting
with the data.

Results from a unit test should be consistent, able to be automated, and
fast. The benefits are that it is easier to find bugs, changes don't
break surrounding code because of testing, it improves code quality, and
can also serve as documentation

The three main parts of a unit test are: arrange, act, and assert

Arrange is where you bring in the parts that need to be tested, and all
of the dependencies or context needed.
Act is where you actually do the testing and you're inputting all of the
parts that are going to be tested.
Assert is where you check the results that came back and make sure that
they are what was expected.